labs/sd-finetuner/requirements.txt
torch>=2.1.0
torchvision>=0.16.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pillow>=10.0.0
pydantic>=2.0.0
numpy>=1.26.0
pytest>=8.0.0
